+++ src/main/java/com/zhenzhang0123/filmsalessystembackend/repository/ShowRepository.java
@@ -8,6 +8,6 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface ShowRepository extends JpaRepository<Show, Integer> {
+public interface ShowRepository extends JpaRepository<Show, Long> {
     List<Show> findByRemainingTicketsGreaterThanAndShowTimeAfter(int remainingTickets, LocalDateTime now);
 }
